+++ test-apps/ui-test-app/src/common/TestAppConfiguration.ts
@@ -11,6 +11,11 @@ export interface TestAppConfiguration {
   bingMapsKey?: string;
   mapBoxKey?: string;
   cesiumIonKey?: string;
+
+  // ArcGis OAuth2 configuration
+  arcGisOnlineClientId?: string;
+  arcGisEnterpriseBaseUrl?: string;
+  arcGisEnterpriseClientId?: string;
 }
 
 export const loggerCategory = "ui-test-app";