+++ ui/appui-react/src/appui-react/statusfields/StatusFieldProps.tsx
@@ -5,8 +5,8 @@
 /** @packageDocumentation
  * @module StatusBar
  */
-
 import { CommonProps } from "@itwin/core-react";
+
 import { StatusBarFieldId } from "../statusbar/StatusBarWidgetControl";
 
 /** Properties for a StatusBar field component
@@ -15,8 +15,12 @@ import { StatusBarFieldId } from "../statusbar/StatusBarWidgetControl";
 export interface StatusFieldProps extends CommonProps {
   /** Indicates whether the StatusBar is in footer mode */
   isInFooterMode: boolean;
-  /** Function called when the widget is opened or closed. */
-  onOpenWidget: (widget: StatusBarFieldId) => void;
-  /** Field Id for open widgets */
-  openWidget: StatusBarFieldId;
+  /** Function called when the widget is opened or closed.
+   * @deprecated In upcoming versions, this will be removed. Field will have the freedom of handling their dialog behavior however they like.
+   */
+  onOpenWidget?: (widget: StatusBarFieldId) => void;
+  /** Field Id for open widgets
+   * @deprecated In upcoming versions, this will be removed. Field will have the freedom of handling their dialog behavior however they like.
+   */
+  openWidget?: StatusBarFieldId;
 }