+++ ui/appui-react/src/appui-react/uistate/AppUiSettings.ts
@@ -25,6 +25,7 @@ export interface InitialAppUiSettings {
   /** @alpha */
   autoCollapseUnpinnedPanels?: boolean;
   animateToolSettings?: boolean;
+  useToolAsToolSettingsLabel?: boolean;
 }
 
 /** These are the UI settings that are stored in the Redux store. They control the color theme, the UI version,
@@ -52,6 +53,7 @@ export class AppUiSettings implements UserSettingsProvider {
   public showWidgetIcon: UiStateEntry<boolean>;
   public autoCollapseUnpinnedPanels: UiStateEntry<boolean>;
   public animateToolSettings: UiStateEntry<boolean>;
+  public useToolAsToolSettingsLabel: UiStateEntry<boolean>;
 
   private setColorTheme = (theme: string) => {
     UiFramework.setColorTheme(theme);
@@ -94,6 +96,11 @@ export class AppUiSettings implements UserSettingsProvider {
       (value: boolean) => UiFramework.setAnimateToolSettings(value), defaults.animateToolSettings);
     this._settings.push(this.animateToolSettings);
 
+    this.useToolAsToolSettingsLabel = new UiStateEntry<boolean>(AppUiSettings._settingNamespace, "UseToolAsToolSettingsLabel",
+      () => UiFramework.useToolAsToolSettingsLabel,
+      (value: boolean) => UiFramework.setUseToolAsToolSettingsLabel(value), defaults.useToolAsToolSettingsLabel);
+    this._settings.push(this.useToolAsToolSettingsLabel);
+
     SyncUiEventDispatcher.onSyncUiEvent.addListener(this.handleSyncUiEvent);
   }
 
@@ -124,6 +131,9 @@ export class AppUiSettings implements UserSettingsProvider {
 
     if (args.eventIds.has("configurableui:set-animate-tool-settings"))
       await this.animateToolSettings.saveSetting(UiFramework.getUiStateStorage());
+
+    if (args.eventIds.has("configurableui:set-use-tool-as-tool-settings-label"))
+      await this.useToolAsToolSettingsLabel.saveSetting(UiFramework.getUiStateStorage());
   };
 
   public async apply(storage: UiStateStorage): Promise<void> {