+++ core/transformer/CHANGELOG.md
@@ -1,6 +1,36 @@
 # Change Log - @itwin/core-transformer
 
-This log was last generated on Fri, 15 Apr 2022 13:49:25 GMT and should not be manually modified.
+This log was last generated on Tue, 21 Jun 2022 18:06:33 GMT and should not be manually modified.
+
+## 3.2.4
+Tue, 21 Jun 2022 18:06:33 GMT
+
+_Version update only_
+
+## 3.2.3
+Fri, 17 Jun 2022 15:18:39 GMT
+
+_Version update only_
+
+## 3.2.2
+Fri, 10 Jun 2022 16:11:36 GMT
+
+_Version update only_
+
+## 3.2.1
+Tue, 07 Jun 2022 15:02:56 GMT
+
+_Version update only_
+
+## 3.2.0
+Fri, 20 May 2022 13:10:54 GMT
+
+### Updates
+
+- always close transformer resumption state db even on errors
+- Add an option to IModelImporter to optimize geometry by identifying geometry parts with only one reference and embedding the part's geometry directly into the referencing element's geometry stream.
+- add transformer resumption API
+- use refactored internal extractChangedIdsFromChangesets
 
 ## 3.1.3
 Fri, 15 Apr 2022 13:49:25 GMT